Random changes to textInput.py

+++ imageMapping.py
@@ -2,7 +2,7 @@ import math
 
 def images(n, imageCount):
   numbers = [i for i in range(1,imageCount+1)]
-  img1 = math.ceil(n / ((imageCount-1)*(imageCount-2)))
+  img1 = int(math.ceil(n / ((imageCount-1)*(imageCount-2))))
   numbers.remove(img1)
 
   img2 = numbers[int(math.ceil((n % ((imageCount-1)*(imageCount-2))) / (imageCount-2))-1)]
